feat(settings): search-only Add Token flow
The contract-address Add Token form is obsolete now that the picker can
register crypto end-to-end via search (Task 9). Replace AddTokenSheet
with a thin wrapper around InstrumentPickerSheet limited to crypto
results, and trim the resolution+confirm orchestration off
CryptoTokenStore.

InstrumentPickerSheet gains a callback initialiser so callers without an
existing Instrument binding (Settings flows) can use it. The render
implementation moves to InstrumentPickerSheetCore to keep both files
under the file-length ceiling; the public InstrumentPickerSheet
dispatches between bound and callback modes.
